Create Node with maximum subtree sum .java
It is a unique concept of generic tree in this we are getting maximum sum subtree in the generic tree and this is the special trick. In this, we getting return type something else and our results are different. generally in our solution whatever we are getting a return in our recursion will be the answer in many situations but in this, we are making a unique approach our answer is different and our return type from recursion is different.