Create Diameter of GenericTree.java
In this question, we are finding the diameter of a generic tree. In the diameter of the generic tree, the distance of the maximum deepest node's last element and the second deepest node's last element +2 is known as the diameter. but it is not the same in every case. if you increase the height of a subtree and it has multiple branches then we have created the logic like that if it is greater than the calculation of deepest node + second-deepest Node +2 then we replace the value of that answer with the maximum distance.